move hour and rank table parsing to ParseUCOSched
